requirements.txt
fastapi
uvicorn[standard]
flask
pandas
docker
typing
stix2
subprocess
confluent-kafka
